% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object-methods.R
\name{summary.hier_gsea_result}
\alias{summary.hier_gsea_result}
\title{Summary for a hierarchy-aware GSEA result}
\usage{
\method{summary}{hier_gsea_result}(object, ...)
}
\arguments{
\item{object}{A \code{hier_gsea_result} object.}

\item{...}{Unused.}
}
\value{
A named list with compact summary metrics.
}
\description{
Summary for a hierarchy-aware GSEA result
}
