% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-package.R
\docType{package}
\name{hierGSEA-package}
\alias{hierGSEA}
\alias{hierGSEA-package}
\title{hierGSEA package}
\description{
Utilities for hierarchy-aware post-processing of GSEA results from
\code{clusterProfiler} and \code{ReactomePA}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://maxull.github.io/hierGSEA/}
  \item \url{https://github.com/maxull/hierGSEA}
  \item Report bugs at \url{https://github.com/maxull/hierGSEA/issues}
}

}
\author{
\strong{Maintainer}: Max Ullrich \email{max@example.com}

}
\keyword{internal}
